Spliting provider/core into two files - one for generation, one for shared code
